requirements.txt
mysql.connector
flask